Holy crap, this ended up being a lot of work
I took the naive approach to rename everything I could, but also leave file extensions as .nix for easier merging
This made it so that some of the struct/class changes needed to be done by hand. I probably got something wrong.
I made a previous patch that additionally loads ".bsd" files in addition to ".nix" files, but this is definitely
a lot more work than I expected. JAH TODO: Need to revisit this patch

In the name of expediency, I got everything patched up so that the `nix` cmd now appears as `bsd`, and hopefully
it also has the same semantics. JAH TODO: Start running the nixos functional tests manually

All told, it would be much better to add a new compilation unit in meson that replaces 's|/nix|/bsd|g' in the nix binary
We also really should be taking advantage of the Nix cache, we're only building from source BECAUSE we changed NIX_STORE_DIR
If the installer worked, but then migrated everything to our mount, that would be cool. Nix has support for path stripping